+++ doc/CHANGELOG.rst
@@ -35,7 +35,8 @@ This changelog *only* contains changes from the *first* pypi release (0.5.4.3) o
 
   - Updated the repr methods on Java classes.
 
-  - ``java.util.List`` completes the contract for ``collections.abc.Sequence``.
+  - ``java.util.List`` completes the contract for ``collections.abc.Sequence``
+    and ``collections.abc.MutableSequence``.
 
   - ``java.util.Collection`` completes the contract for ``collections.abc.Collection``.